detach Item Class from MenuSite and Refactor Menu Class
