+++ HolubLife/src/com/holub/ui/MenuItem.java
@@ -1,0 +1,8 @@
+package com.holub.ui;
+
+import com.holub.life.controller.Universe;
+import com.holub.life.view.UniverseView;
+
+public interface MenuItem {
+    public void setup();
+}
