+++ HolubLife/test/com/holub/life/ClockTest.java
@@ -1,28 +1,28 @@
-package com.holub.life;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ClockTest {
-
-    @Test
-    void instance() {
-    }
-
-    @Test
-    void startTicking() {
-    }
-
-    @Test
-    void stop() {
-    }
-
-    @Test
-    void addClockListener() {
-    }
-
-    @Test
-    void tick() {
-    }
-}
+//package com.holub.life;
+//
+//import org.junit.jupiter.api.Test;
+//
+//import static org.junit.jupiter.api.Assertions.*;
+//
+//class ClockTest {
+//
+//    @Test
+//    void instance() {
+//    }
+//
+//    @Test
+//    void startTicking() {
+//    }
+//
+//    @Test
+//    void stop() {
+//    }
+//
+//    @Test
+//    void addClockListener() {
+//    }
+//
+//    @Test
+//    void tick() {
+//    }
+//}
